refactor(database/repos): move whitelist check to new repo
