+++ packages/database/src/repositories/players.ts
@@ -19,7 +19,6 @@ import {
 	warns,
 	kicks,
 	playerNotes,
-	whitelistedIdentifers,
 } from '../schema';
 import type * as schema from '../schema';
 import type {
@@ -415,30 +414,6 @@ class PlayersRepository {
 			issuer: adminId,
 			issuedAt: new Date(),
 		});
-	}
-
-	async isAnyIdentifierWhitelisted(
-		identifiers: PlayerIdentifiers,
-	): Promise<boolean> {
-		const conditions = [eq(whitelistedIdentifers.value, identifiers.license)];
-
-		if (identifiers.fivem) {
-			conditions.push(eq(whitelistedIdentifers.value, identifiers.fivem));
-		}
-		if (identifiers.discord) {
-			conditions.push(eq(whitelistedIdentifers.value, identifiers.discord));
-		}
-		if (identifiers.steam) {
-			conditions.push(eq(whitelistedIdentifers.value, identifiers.steam));
-		}
-
-		const result = await this.db
-			.select({ id: whitelistedIdentifers.id })
-			.from(whitelistedIdentifers)
-			.where(or(...conditions))
-			.limit(1);
-
-		return result.length > 0;
 	}
 }
